+++ Components/Screens/AppHomeScreen.js
@@ -2,6 +2,7 @@ import React from "react";
 import { View, Text, StyleSheet } from "react-native";
 import { TouchableHighlight } from "react-native-gesture-handler";
 import theme from '../../theme';
+import {styles} from './styleAppHomeScreen';
 
 
 const AppHomeScreen = (props) => {
@@ -42,50 +43,5 @@ const AppHomeScreen = (props) => {
       </View>
     );
 }
-
-const styles = StyleSheet.create({
-  headerContainer: {
-    borderBottomColor: theme.primary,
-    borderBottomWidth: StyleSheet.hairlineWidth,
-    marginBottom: 30
-  },
-  header: {
-    fontSize: 50,
-    fontFamily: theme.textMain
-  },
-  greeting: {
-    fontSize: 30,
-    fontFamily: theme.textMain,
-    margin: 10,
-    marginBottom: 30
-  },
-  button: {
-    color: theme.accentOne,
-    fontSize: 30,
-    fontFamily: theme.textTwo,
-    textAlign: "center",
-    paddingTop: 25,
-  },
-  container: {
-    backgroundColor: theme.accentOne,
-    alignItems: "center",
-    justifyContent: "center",
-    height: "100%"
-  },
-  routes: {
-    flexDirection: "column",
-    backgroundColor: theme.primary,
-    width: "80%",
-    height: "15%",
-    justifyContent: "space-around",
-    margin: 10,
-    borderRadius: 50,
-  },
-  touchExpander: {
-		height: '100%',
-		borderRadius: 50,
-		width: '100%'
-	}
-});
 
 export default AppHomeScreen
